⚡️ New navigator that keeps state of pages
