+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "dart.flutterSdkPath": "D:\\Leo\\Flutter\\insta_code\\.fvm\\flutter_sdk"
+}
