+++ lib/main.dart
@@ -1,8 +1,7 @@
-import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_core/firebase_core.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:insta_code/widgets/nav_bar.dart';
+import 'package:insta_code/utils/firebase_methods.dart';
+import 'package:insta_code/widgets/navigation.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -20,15 +19,7 @@ class MyApp extends StatelessWidget {
         accentColor: Colors.black,
         visualDensity: VisualDensity.adaptivePlatformDensity,
       ),
-      home: const NavBar(),
+      home: Navigation(),
     );
   }
-}
-
-void initFirebase() {
-  final String host = defaultTargetPlatform == TargetPlatform.android
-      ? '10.0.2.2:8080'
-      : 'localhost:8080';
-
-  FirebaseFirestore.instance.settings = Settings(host: host, sslEnabled: false);
 }
