+++ lib/utils/tab_items.dart
@@ -1,0 +1,6 @@
+const List<String> pages = [
+  "Home",
+  "Liked",
+  "Create",
+  "Profile",
+];
